Extract interface to file
